Add CSR MIMPID value configuration, and set to '1'
Add a config for the MIMPID CSR and set this to 1, as we've now
fixed a fairly annyoing issue for Linux where the FP dirty state
wasn't correctly set.

This show allow the setup code to patch the fp-save code to deal
with whether it should always save the fpu state.

See commit 2700d14471963909923bdcfa1b013b4d1b30e567 for the fix
for the FS status (i think)

Signed-off-by: Ben Dooks <ben.dooks@codethink.co.uk>